FEAT: separando template products e adicionando campo marca

+++ imports/views.py
@@ -2,7 +2,8 @@ from operator import contains
 
 from django.contrib import messages
 from django.contrib.messages import constants
-from django.shortcuts import redirect, render
+from django.shortcuts import (get_list_or_404, get_object_or_404, redirect,
+                              render)
 
 from .forms import ProdutoForm
 from .models import Produto
@@ -38,6 +39,15 @@ def products(request):
 
 def delete_product(request, product_id):
     product = Produto.objects.filter(id=product_id)
-    product.delete()
-    messages.add_message(request, constants.WARNING, 'Produto deletado')
+
+    if request.method == 'GET':
+        return redirect('/products')
+
+    if request.method == 'POST':
+        product.delete()
+        messages.add_message(request, constants.WARNING, 'Produto deletado')
+        return redirect('/products')
+
+
+def edit_product(request, product_id):
     return redirect('/products')
